requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
line-bot-sdk==3.14.5
deep-translator==1.11.4
